refactor and follow ups
