+++ src/components/src/effects/effect-configurator.tsx
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 // Copyright contributors to the kepler.gl project
 
-import React, {useCallback, useMemo, useState} from 'react';
+import React, {useCallback, useMemo} from 'react';
 import styled from 'styled-components';
 import {injectIntl, IntlShape} from 'react-intl';
 
@@ -10,10 +10,11 @@ import {isNumber} from '@kepler.gl/utils';
 import {Effect, EffectUpdateProps} from '@kepler.gl/types';
 
 import RangeSliderFactory from '../common/range-slider';
-import {ArrowDownSmall, VertThreeDots} from '../common/icons';
+import {ArrowDownSmall} from '../common/icons';
 import {Tooltip} from '../common/styled-components';
 import EffectTimeConfiguratorFactory from './effect-time-configurator';
 import CompactColorPicker from './compact-color-picker';
+import SurfaceFogElevationSectionFactory, {ELEVATION_SECTION_TYPE} from './surface-fog-section';
 
 export type EffectConfiguratorProps = {
   effect: Effect;
@@ -190,56 +191,6 @@ const StyledHintIcon = styled.span`
   }
 `;
 
-const StyledCollapsibleSectionHeader = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  cursor: pointer;
-  margin-bottom: 4px;
-  &:hover {
-    .effect-section__title {
-      color: ${props => props.theme.textColorHl};
-    }
-    .effect-section__dots {
-      color: ${props => props.theme.textColorHl};
-    }
-  }
-`;
-
-const StyledElevationSectionTitle = styled.div`
-  border-left: ${props => props.theme.layerConfigGroupLabelBorderLeft} solid
-    ${props => props.theme.labelColor};
-  line-height: 12px;
-  padding-left: ${props => props.theme.layerConfigGroupLabelPadding};
-  font-size: ${props => props.theme.layerConfigGroupLabelLabelFontSize};
-  color: ${props => props.theme.textColor};
-  font-weight: 500;
-  letter-spacing: 0.2px;
-  text-transform: capitalize;
-`;
-
-const StyledDotsWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  color: ${props => props.theme.textColor};
-`;
-
-const StyledCollapsibleContent = styled.div<{$collapsed: boolean}>`
-  overflow: hidden;
-  max-height: ${props => (props.$collapsed ? '0' : '200px')};
-  transition: max-height 0.3s ease-out;
-`;
-
-const StyledDisabledSlider = styled.div<{$disabled: boolean}>`
-  ${props =>
-    props.$disabled
-      ? `
-    opacity: 0.3;
-    pointer-events: none;
-  `
-      : ''}
-`;
-
 const StyledColorCheckboxRow = styled.div`
   display: flex;
   align-items: center;
@@ -258,17 +209,23 @@ type EffectParameterDescriptionFlattened = {
   name: string;
   type?: 'number' | 'array' | 'color' | 'checkbox';
   label?: string | false | (string | false)[];
+  tooltip?: string;
   min: number;
   max: number;
   defaultValue?: number | number[] | boolean;
   index?: number;
 };
 
-EffectConfiguratorFactory.deps = [RangeSliderFactory, EffectTimeConfiguratorFactory];
+EffectConfiguratorFactory.deps = [
+  RangeSliderFactory,
+  EffectTimeConfiguratorFactory,
+  SurfaceFogElevationSectionFactory
+];
 
 export default function EffectConfiguratorFactory(
   RangeSlider: ReturnType<typeof RangeSliderFactory>,
-  EffectTimeConfigurator: ReturnType<typeof EffectTimeConfiguratorFactory>
+  EffectTimeConfigurator: ReturnType<typeof EffectTimeConfiguratorFactory>,
+  SurfaceFogElevationSection: ReturnType<typeof SurfaceFogElevationSectionFactory>
 ): React.FC<EffectConfiguratorProps> {
   const ShadowEffectConfigurator: React.FC<EffectConfiguratorProps> = ({
     effect,
@@ -393,6 +350,8 @@ export default function EffectConfiguratorFactory(
   };
   const defaultUniforms = {};
 
+  const ELEVATION_EXTRA_PARAMS = new Set(['heightEnd', 'animateHeight', 'linearEasing']);
+
   const PostProcessingEffectConfigurator: React.FC<EffectConfiguratorProps> = ({
     effect,
     updateEffectConfig
@@ -401,16 +360,16 @@ export default function EffectConfiguratorFactory(
     const parameterDescriptions = effect.getParameterDescriptions();
     const {parameters, id} = effect;
 
-    const isAnimateHeight = Boolean(parameters.animateHeight);
+    const hasElevationSection = parameterDescriptions.some(d => d.name === 'animateHeight');
 
     const flatParameterDescriptions = useMemo(() => {
       return parameterDescriptions.reduce((acc, description) => {
-        if (
-          description.name === 'heightEnd' ||
-          description.name === 'animateHeight' ||
-          description.name === 'linearEasing'
-        )
+        if (hasElevationSection && ELEVATION_EXTRA_PARAMS.has(description.name)) return acc;
+
+        if (hasElevationSection && description.name === 'height') {
+          acc.push({...description, type: ELEVATION_SECTION_TYPE as any});
           return acc;
+        }
 
         if (description.type === 'color') {
           acc.push(description);
@@ -430,14 +389,14 @@ export default function EffectConfiguratorFactory(
 
         return acc;
       }, [] as EffectParameterDescriptionFlattened[]);
-    }, [parameterDescriptions]);
-
-    const heightEndDesc = parameterDescriptions.find(d => d.name === 'heightEnd');
-    const hasAnimateHeight = parameterDescriptions.some(d => d.name === 'animateHeight');
-    const hasLinearEasing = parameterDescriptions.some(d => d.name === 'linearEasing');
+    }, [parameterDescriptions, hasElevationSection]);
 
     const controls = useMemo(() => {
       return flatParameterDescriptions.map(desc => {
+        if ((desc.type as string) === ELEVATION_SECTION_TYPE) {
+          return {isElevationSection: true as const};
+        }
+
         if (desc.type === 'color') {
           return {
             isColor: true as const,
@@ -450,76 +409,19 @@ export default function EffectConfiguratorFactory(
         }
 
         if (desc.type === 'checkbox') {
-          const label =
-            desc.name === 'pattern'
-              ? 'Pattern'
-              : typeof desc.label === 'string'
-              ? desc.label
-              : desc.name;
-          const tooltip =
-            desc.name === 'pattern'
-              ? 'Adds a noise pattern to the fog for a more natural, volumetric look.'
-              : undefined;
+          const label = typeof desc.label === 'string' ? desc.label : desc.name;
           return {
             isCheckbox: true as const,
             label,
             paramName: desc.name,
             checked: Boolean(parameters[desc.name]),
-            tooltip,
+            tooltip: desc.tooltip,
             onChange: () =>
               updateEffectConfig(null, id, {parameters: {[desc.name]: !parameters[desc.name]}})
           };
         }
 
         const paramName = desc.name;
-
-        // Elevation section: main slider + collapsible animate checkbox + end slider
-        if (paramName === 'height' && hasAnimateHeight && heightEndDesc) {
-          const rangeMin = desc.min ?? -200;
-          const rangeMax = desc.max ?? 3000;
-          const rangeSpan = rangeMax - rangeMin;
-          const step = rangeSpan > 10 ? 1 : 0.001;
-
-          const endMin = heightEndDesc.min ?? rangeMin;
-          const endMax = heightEndDesc.max ?? rangeMax;
-          const endSpan = endMax - endMin;
-          const endStep = endSpan > 10 ? 1 : 0.001;
-
-          return {
-            isElevationSection: true as const,
-            label: desc.label || desc.name,
-            value1: parameters.height ?? desc.defaultValue ?? 0,
-            value0: rangeMin,
-            range: [rangeMin, rangeMax] as [number, number],
-            step,
-            onChange: (newValue: number[], event?: Event | null) => {
-              updateEffectConfig(event, id, {parameters: {height: newValue[1]}});
-            },
-            animateChecked: isAnimateHeight,
-            onAnimateChange: () => {
-              const enabling = !parameters.animateHeight;
-              const currentEnd = parameters.heightEnd ?? heightEndDesc.max ?? rangeMax;
-              const currentHeight = parameters.height ?? desc.defaultValue ?? 0;
-              const updatedParams: Record<string, any> = {animateHeight: enabling};
-              if (enabling && Math.abs(currentEnd - currentHeight) < 10) {
-                updatedParams.heightEnd = rangeMax;
-              }
-              updateEffectConfig(null, id, {parameters: updatedParams});
-            },
-            endValue1: parameters.heightEnd ?? heightEndDesc.defaultValue ?? endMax,
-            endValue0: endMin,
-            endRange: [endMin, endMax] as [number, number],
-            endStep,
-            onEndChange: (newValue: number[], event?: Event | null) => {
-              updateEffectConfig(event, id, {parameters: {heightEnd: newValue[1]}});
-            },
-            linearEasingChecked: Boolean(parameters.linearEasing),
-            hasLinearEasing,
-            onLinearEasingChange: () => {
-              updateEffectConfig(null, id, {parameters: {linearEasing: !parameters.linearEasing}});
-            }
-          };
-        }
 
         const rawUniform = uniforms[desc.name];
         if (rawUniform && rawUniform.private) {
@@ -583,7 +485,6 @@ export default function EffectConfiguratorFactory(
           };
         }
         // Parameter defined in effect description but not in shader propTypes
-        // (e.g. user-facing props that the effect maps to different GLSL uniforms)
         else if (desc.min !== undefined && desc.max !== undefined) {
           const rangeSpan = desc.max - desc.min;
           const step = rangeSpan > 10 ? 1 : 0.001;
@@ -602,22 +503,7 @@ export default function EffectConfiguratorFactory(
         // ignore everything else for now
         return null;
       });
-    }, [
-      flatParameterDescriptions,
-      id,
-      parameters,
-      updateEffectConfig,
-      uniforms,
-      isAnimateHeight,
-      hasAnimateHeight,
-      heightEndDesc,
-      hasLinearEasing
-    ]);
-
-    const [elevationExpanded, setElevationExpanded] = useState(false);
-    const onToggleElevationExpanded = useCallback(() => {
-      setElevationExpanded(prev => !prev);
-    }, []);
+    }, [flatParameterDescriptions, id, parameters, updateEffectConfig, uniforms]);
 
     return (
       <StyledEffectConfigurator key={effect.id}>
@@ -700,101 +586,6 @@ export default function EffectConfiguratorFactory(
               continue;
             }
 
-            if ('isElevationSection' in control) {
-              const tooltipId = `effect-elevation-hint-${effect.id}`;
-              elements.push(
-                <RegularOuterWrapper key={`${effect.id}-${i}`}>
-                  <StyledCollapsibleSectionHeader onClick={onToggleElevationExpanded}>
-                    <StyledElevationSectionTitle className="effect-section__title">
-                      Elevation
-                    </StyledElevationSectionTitle>
-                    <StyledDotsWrapper className="effect-section__dots">
-                      <VertThreeDots height="18px" />
-                    </StyledDotsWrapper>
-                  </StyledCollapsibleSectionHeader>
-                  <RegularSliderWrapper>
-                    <RangeSlider
-                      {...COMMON_SLIDER_PROPS}
-                      value1={control.value1 as number}
-                      value0={control.value0 as number}
-                      range={control.range}
-                      step={control.step}
-                      onChange={control.onChange}
-                    />
-                  </RegularSliderWrapper>
-                  <StyledCollapsibleContent $collapsed={!elevationExpanded}>
-                    <RegularOuterWrapper style={{marginTop: 8}}>
-                      <StyledCheckboxWrapper onClick={control.onAnimateChange}>
-                        <StyledCheckbox
-                          type="checkbox"
-                          checked={control.animateChecked}
-                          onChange={control.onAnimateChange}
-                          onClick={e => e.stopPropagation()}
-                        />
-                        <StyledCheckboxLabel>Animate elevation</StyledCheckboxLabel>
-                        <StyledHintIcon
-                          data-tip
-                          data-for={tooltipId}
-                          onClick={e => e.stopPropagation()}
-                        >
-                          ?
-                        </StyledHintIcon>
-                        <StyledHintTooltip id={tooltipId} effect="solid" delayShow={200}>
-                          Animates elevation from start to end value during video export preview and
-                          recording.
-                        </StyledHintTooltip>
-                      </StyledCheckboxWrapper>
-                    </RegularOuterWrapper>
-                    <StyledDisabledSlider $disabled={!control.animateChecked}>
-                      <RegularSectionTitleWrapper>End elevation</RegularSectionTitleWrapper>
-                      <RegularSliderWrapper>
-                        <RangeSlider
-                          {...COMMON_SLIDER_PROPS}
-                          value1={control.endValue1 as number}
-                          value0={control.endValue0 as number}
-                          range={control.endRange}
-                          step={control.endStep}
-                          onChange={control.onEndChange}
-                        />
-                      </RegularSliderWrapper>
-                    </StyledDisabledSlider>
-                    {control.hasLinearEasing ? (
-                      <StyledDisabledSlider $disabled={!control.animateChecked}>
-                        <RegularOuterWrapper style={{marginTop: 4}}>
-                          <StyledCheckboxWrapper onClick={control.onLinearEasingChange}>
-                            <StyledCheckbox
-                              type="checkbox"
-                              checked={control.linearEasingChecked}
-                              onChange={control.onLinearEasingChange}
-                              onClick={e => e.stopPropagation()}
-                            />
-                            <StyledCheckboxLabel>Linear easing</StyledCheckboxLabel>
-                            <StyledHintIcon
-                              data-tip
-                              data-for={`effect-linear-hint-${effect.id}`}
-                              onClick={e => e.stopPropagation()}
-                            >
-                              ?
-                            </StyledHintIcon>
-                            <StyledHintTooltip
-                              id={`effect-linear-hint-${effect.id}`}
-                              effect="solid"
-                              delayShow={200}
-                            >
-                              Uses constant speed instead of smooth ease-in / ease-out during
-                              elevation animation.
-                            </StyledHintTooltip>
-                          </StyledCheckboxWrapper>
-                        </RegularOuterWrapper>
-                      </StyledDisabledSlider>
-                    ) : null}
-                  </StyledCollapsibleContent>
-                </RegularOuterWrapper>
-              );
-              i++;
-              continue;
-            }
-
             if ('isCheckbox' in control) {
               elements.push(
                 <RegularOuterWrapper key={`${effect.id}-${i}`}>
@@ -815,6 +606,18 @@ export default function EffectConfiguratorFactory(
               continue;
             }
 
+            if ('isElevationSection' in control) {
+              elements.push(
+                <SurfaceFogElevationSection
+                  key={`${effect.id}-elevation`}
+                  effect={effect}
+                  updateEffectConfig={updateEffectConfig}
+                />
+              );
+              i++;
+              continue;
+            }
+
             elements.push(
               <RegularOuterWrapper key={`${effect.id}-${i}`}>
                 {control.label ? (
@@ -827,6 +630,7 @@ export default function EffectConfiguratorFactory(
             );
             i++;
           }
+
           return elements;
         })()}
       </StyledEffectConfigurator>
