+++ src/components/src/modals/export-video-modal.tsx
@@ -4,12 +4,7 @@
 import React, {useCallback, useEffect, useMemo, useRef, useState} from 'react';
 import styled, {ThemeProvider, useTheme} from 'styled-components';
 
-import {
-  DEFAULT_MAPBOX_API_URL,
-  NO_MAP_ID,
-  EMPTY_MAPBOX_STYLE,
-  SURFACE_FOG_TYPE
-} from '@kepler.gl/constants';
+import {DEFAULT_MAPBOX_API_URL, NO_MAP_ID, EMPTY_MAPBOX_STYLE} from '@kepler.gl/constants';
 import {FormattedMessage} from '@kepler.gl/localization';
 import {Viewport, ExportVideo, Effect} from '@kepler.gl/types';
 import {
@@ -34,6 +29,7 @@ import {
   getTimeRangeFilterKeyframes,
   getAnimatableFilters
 } from './hubble-utils';
+import {useFogHeightAnimation} from './fog-height-animation';
 
 type HubbleModule = {
   ExportVideoPanelContainer: React.ComponentType<any>;
@@ -296,21 +292,6 @@ const ExportVideoModalFactory = () => {
       (visState.effects || []).map((effect: Effect) => effect.clone())
     );
 
-    const surfaceFogEffect = useMemo(
-      () => videoEffects.find((e: Effect) => e.type === SURFACE_FOG_TYPE && e.isEnabled),
-      [videoEffects]
-    );
-
-    const fogHeightAnim = useMemo(() => {
-      if (!surfaceFogEffect?.parameters?.animateHeight) return null;
-      const params = surfaceFogEffect.parameters;
-      return {
-        start: params.height as number,
-        end: params.heightEnd as number,
-        linear: Boolean(params.linearEasing)
-      };
-    }, [surfaceFogEffect]);
-
     const [videoConfiguration, setVideoConfiguration] = useState<VideoConfiguration>({
       ...exportVideo
     });
@@ -319,69 +300,9 @@ const ExportVideoModalFactory = () => {
       []
     );
 
-    const fogAnimRef = useRef<{
-      active: boolean;
-      range: [number, number];
-      originalHeight: number;
-      linear: boolean;
-    }>({active: false, range: [0, 100], originalHeight: 50, linear: false});
-
     const hubbleContainerRef = useRef<any>(null);
 
-    // Sync fogAnimRef config with effect parameters (don't start playing yet)
-    useEffect(() => {
-      if (fogHeightAnim) {
-        fogAnimRef.current.active = true;
-        fogAnimRef.current.range = [fogHeightAnim.start, fogHeightAnim.end];
-        fogAnimRef.current.originalHeight = fogHeightAnim.start;
-        fogAnimRef.current.linear = fogHeightAnim.linear;
-      } else {
-        if (fogAnimRef.current.active && surfaceFogEffect?.deckEffect) {
-          surfaceFogEffect.deckEffect.setProps?.({height: fogAnimRef.current.originalHeight});
-        }
-        fogAnimRef.current.active = false;
-      }
-    }, [fogHeightAnim, surfaceFogEffect]);
-
-    // Patch DeckSurfaceFogEffect.postRender to drive height animation per-frame.
-    // Reads the Hubble adapter's videoCapture.timeMs to derive progress,
-    // so our animation stays perfectly in sync with Hubble's own timeline.
-    useEffect(() => {
-      const de = surfaceFogEffect?.deckEffect;
-      if (!de) return;
-      const originalPostRender = de.postRender.bind(de);
-      de.postRender = function (this: any, params: any) {
-        const ref = fogAnimRef.current;
-        if (ref.active) {
-          const hubbleState = hubbleContainerRef.current?.state;
-          const isPlaying = hubbleState?.previewing || hubbleState?.rendering;
-
-          if (isPlaying) {
-            const adapter = hubbleState?.adapter;
-            const vc = adapter?.videoCapture;
-            const tc = vc?.timecode;
-            if (
-              vc &&
-              tc &&
-              tc.duration > 0 &&
-              Number.isFinite(vc.timeMs) &&
-              Number.isFinite(tc.start)
-            ) {
-              const progress = Math.min(Math.max((vc.timeMs - tc.start) / tc.duration, 0), 1);
-              const t = ref.linear ? progress : progress * progress * (3 - 2 * progress);
-              const [s, e] = ref.range;
-              this.props.height = s + (e - s) * t;
-            }
-          } else {
-            this.props.height = ref.originalHeight;
-          }
-        }
-        return originalPostRender(params);
-      };
-      return () => {
-        de.postRender = originalPostRender;
-      };
-    }, [surfaceFogEffect]);
+    useFogHeightAnimation(videoEffects, hubbleContainerRef);
 
     const deckEffects = useMemo(() => {
       if (videoEffects.length === 0) return [];
